+++ src/extension.ts
@@ -89,7 +89,7 @@ async function activate(context: vscode.ExtensionContext) {
     const CONNECT_SEQUENCES_DOC_URL = "https://calva.io/connect-sequences/";
     const CALVA_DOCS_URL = "https://calva.io/";
     const VIEWED_CALVA_DOCS = "viewedCalvaDocs";
-
+    const VIEWED_SHORTCUT_CHANGE_MSG = "viewedShortcutsChangeMessage";
     if (customCljsRepl && replConnectSequences.length == 0) {
         chan.appendLine("Old customCljsRepl settings detected.");
         vscode.window.showErrorMessage("Old customCljsRepl settings detected. You need to specify it using the new calva.customConnectSequence setting. See the Calva user documentation for instructions.", ...[BUTTON_GOTO_DOC, BUTTON_OK])
@@ -190,11 +190,6 @@ async function activate(context: vscode.ExtensionContext) {
             });
     }))
 
-    // Temporary command to teach new default keyboard shortcut chording key
-    context.subscriptions.push(vscode.commands.registerCommand('calva.tellAboutNewChordingKey', () => {
-        vscode.window.showInformationMessage(`The ”Calva key” has changed. It is now: ctrl+alt+c`);
-    }));
-
     // Initial set of the provided contexts
     outputWindow.setContextForOutputWindowActive(false);
     vscode.commands.executeCommand("setContext", "calva:launching", false);
@@ -266,13 +261,14 @@ async function activate(context: vscode.ExtensionContext) {
 
     greetings.activationGreetings(chan);
 
-    if (!context.globalState.get(VIEWED_CALVA_DOCS)) {
-        vscode.window.showInformationMessage("Calva is activated. Please use the command **Calva: Open Documentation** to visit calva.io for instructions on how to connect Calva to the REPL. (This message will keep showing on start of Calva until you've used the command.)", ...[BUTTON_OK])
+    if (!context.globalState.get(VIEWED_SHORTCUT_CHANGE_MSG)) {
+        vscode.window.showInformationMessage("Dear Calva user, some default keyboard shortcuts have changed. See calva.io/paredit and calva.io/custom-commands for the current defaults.", ...[BUTTON_OK])
             .then(v => {
                 if (v == BUTTON_OK) {
-                    state.analytics().logEvent("LifeCycle", "Greetings dismissed")
+                    context.globalState.update(VIEWED_SHORTCUT_CHANGE_MSG, true);
+                    state.analytics().logEvent("LifeCycle", "Shortcut change message dismissed")
                 }
-            })
+            });
     }