+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes to Calva.
 
 ## [Unreleased]
 - [Add command for copying jack-in command to clipboard](https://github.com/BetterThanTomorrow/calva/pull/995)
+- [Change default shortcuts for Paredit forward/backward sexp, expand/shrink selection, and for slurping and barfing](https://github.com/BetterThanTomorrow/calva/issues/950)
+- [Add Custom Commands variables for current form and more](https://github.com/BetterThanTomorrow/calva/issues/986)
 
 ## [2.0.156] - 2021-01-28
 - Fix: [Debug instrumentation decoration not working correctly anymore on Windows](https://github.com/BetterThanTomorrow/calva/issues/969)